[REMOVE] 사용하지 않는 서비스 제거
