+++ src/main/java/com/ll/playon/domain/game/scheduler/scheduler/GameScheduler.java
@@ -1,6 +1,8 @@
 package com.ll.playon.domain.game.scheduler.scheduler;
 
-import com.ll.playon.domain.game.scheduler.WeeklyGameService;
+import com.ll.playon.domain.game.game.entity.WeeklyPopularGame;
+import com.ll.playon.domain.game.scheduler.repository.WeeklyGameRepository;
+import com.ll.playon.domain.party.party.repository.PartyRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
@@ -8,11 +10,14 @@ import org.springframework.stereotype.Component;
 import java.time.DayOfWeek;
 import java.time.LocalDate;
 import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
 
 @RequiredArgsConstructor
 @Component
 public class GameScheduler {
-    private final WeeklyGameService weeklyGameService;
+    private final PartyRepository partyRepository;
+    private final WeeklyGameRepository weeklyGameRepository;
 
     @Scheduled(cron = "0 0 0 * * MON") // 매주 월요일 00:00
 //    @Scheduled(cron = "0 * * * * *")
@@ -21,6 +26,22 @@ public class GameScheduler {
         LocalDateTime fromDate = weekStart.minusWeeks(1).atStartOfDay(); // 지난주
         LocalDateTime toDate = weekStart.atStartOfDay(); // 이번주
 
-        weeklyGameService.updatePopularGames(fromDate, toDate, weekStart);
+        updatePopularGames(fromDate, toDate, weekStart);
+    }
+
+    public void updatePopularGames(LocalDateTime fromDate, LocalDateTime toDate, LocalDate weekStart) {
+        final int limit = 3;
+
+        List<Map<String, Object>> result = partyRepository.findTopGamesByPartyLastWeek(fromDate, toDate, limit);
+
+        List<WeeklyPopularGame> list = result.stream()
+                .map(row -> WeeklyPopularGame.builder()
+                        .gameId(((Number) row.get("gameId")).longValue())
+                        .playCount(((Number) row.get("playCount")).longValue())
+                        .weekStartDate(weekStart)
+                        .build())
+                .toList();
+
+        weeklyGameRepository.saveAll(list);
     }
 }
